WIP: Use Manifold's tessellator
