+++ src/geometry/cgal/cgalutils.h
@@ -135,4 +135,7 @@ bool corefineAndComputeDifference(TriangleMesh& lhs, TriangleMesh& rhs, Triangle
 template <typename K>
 void convertNefPolyhedronToTriangleMesh(const CGAL::Nef_polyhedron_3<K>& nef, CGAL::Surface_mesh<CGAL::Point_3<K>>& mesh);
 void cleanupMesh(CGAL_HybridMesh& mesh, bool is_corefinement_result);
+
+std::unique_ptr<PolySet> createTriangulatedPolySetFromPolygon2d(const Polygon2d& polygon2d);
+
 } // namespace CGALUtils
