+++ src/geometry/manifold/manifoldutils.h
@@ -23,4 +23,6 @@ namespace ManifoldUtils {
   // FIXME: This shouldn't return const, but it does due to internal implementation details.
   std::shared_ptr<const Geometry> applyMinkowskiManifold(const Geometry::Geometries& children);
 #endif
+
+  std::unique_ptr<PolySet> createTriangulatedPolySetFromPolygon2d(const Polygon2d& polygon2d);
 };
